Makefile: add -grecord-gcc-switches if using -mfentry
Commit a2546fae [ftrace: Add -mfentry to Makefile on function tracer]
adds support for using -mfentry when possible, whenever the ftrace
CONFIG_FUNCTION_TRACER option is enabled.

This unfortunately causes bogus dwarf debuginfo to be generated:

  gcc Bug 54793 - the location of a formal_parameter is not started from a
                  function entry with -mfentry

  http://gcc.gnu.org/bugzilla/show_bug.cgi?id=54793

This can be fixed by applying the gcc patch above, which will be
present in gcc-4.8, but in the meantime, the problem is currently
being worked around in different ways by various tools.  One of those
tools that we support in Yocto, Systemtap, requires the DW_AT_producer
string contain -mfentry, which means that it's assuming
record-gcc-switches is also used:

  SystemTap Bug 15123 - workaround for bad debuginfo for -mfentry
  http://sourceware.org/bugzilla/show_bug.cgi?id=15123

The lack of -grecord-gcc-switches and therefore a missing -mfentry
string in Yocto leads to the following Yocto bug:

  Bug 4099 - Crosstap script check fails
  https://bugzilla.yoctoproject.org/show_bug.cgi?id=4099

Which is what this patch intends to fix.

It only affects architectures that HAVE_FENTRY when
CONFIG_FUNCTION_TRACER is turned on and in any case is just adding to
the debuginfo, so there doesn't seem to be much risk to adding it.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>