Fix for bug 1146 in which the camera had the wrong integration period when running under the 3.8.4 linux kernel.
The bug was caused by a workaround for an issue with the Omap 3630 pll dividers which get loaded with reset values after their PWRDN bits are set - The workaround was implemented in a previous version of the kernel in which the Common Clock Framework was not used, and had not been updated since the kernel transitioned to use the Common Clock Framework. Consequently, the workaround was not correctly restoring the pll divider values when the clock was enabled (except for the first time). In the context of Alinos and the Aptiva mt9v034 camera, this meant that CAM_MCLK was being provided with a clock of around 33Mhz, rather than the correct clock frequency of 28.8Mhz.