Adding FocalTech FT5306 touch panel driver
Also added the touch panel to the Fili-DL device tree.

Note that the driver, while functional, requires some further development
when programming data is acquired for the device. In the mean time
power management is not supported and interrupts will be generated (but
ignored) when the device is not open.

Refs: #1776