[FIX] Fix kernel lock-up during camera capture
Applied the patches from the CameraModule repository to fix the bug causing kernel lock-ups sometimes during camera capture.
The main part of the fix is moving the manipulation of the RESET GPIO from the poweron function to the probe function.
This results in the camera not being reset every session.

Refs: #2052